+++ examples/paperclip.rs
@@ -7,7 +7,9 @@
 // For example: cargo run --example paperclip --features paperclip_compat
 
 use actix_web::{middleware, App, HttpServer};
-use actix_web_middleware_keycloak_auth::{DecodingKey, KeycloakAuth, Role, StandardKeycloakClaims};
+use actix_web_middleware_keycloak_auth::{
+    AlwaysReturnPolicy, DecodingKey, KeycloakAuth, Role, StandardKeycloakClaims,
+};
 use paperclip::actix::{api_v2_operation, web, OpenApiExt};
 
 const KEYCLOAK_PK: &str = "-----BEGIN PUBLIC KEY-----
@@ -59,6 +61,7 @@ async fn main() -> std::io::Result<()> {
     HttpServer::new(|| {
         let keycloak_auth = KeycloakAuth {
             detailed_responses: true,
+            passthrough_policy: AlwaysReturnPolicy,
             keycloak_oid_public_key: DecodingKey::from_rsa_pem(KEYCLOAK_PK.as_bytes()).unwrap(),
             required_roles: vec![Role::Realm {
                 role: "test".to_owned(),