+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## unreleased
 
 - support Actix Web 4.0.0-beta.10
+- add passthrough policy setting to allow auth to be optional
+- add a `KeycloakAuth::default_with_pk()` helper function to initialize the middleware with default settings
+- improve extractors error types
+- expose a pure function to extract custom JWT claims from an Actix Web request
 
 ## v0.4.0-beta.1